requirements.txt
asyncpraw==7.4.0
psycopg2-binary==2.9.1
discord_py_slash_command==3.0.1
discord==1.7.3
